+++ src/thameslib/Site.cc
@@ -7,8 +7,7 @@
 #include "Site.h"
 #include "ChemicalSystem.h"
 
-Site::Site ()
-{
+Site::Site() {
   x_ = y_ = z_ = 0;
   id_ = 0;
   nb_.clear();
@@ -17,47 +16,40 @@ Site::Site ()
   expstrain_ = 0.0;
 }
 
-Site::Site (unsigned int xp,
-            unsigned int yp,
-            unsigned int zp,
-            unsigned int xs,
-            unsigned int ys,
-            unsigned int zs,
-            unsigned int neigh,
-            ChemicalSystem *csys,
-            const bool verbose)
-{    
-    x_ = y_ = z_ = 0;
-    id_ = 0;
-    nb_.clear();
-    stressFreeVolume_ = trueVolume_ = 1.0;
-    damage_ = false;
-    x_ = xp;
-    y_ = yp;
-    z_ = zp;
+Site::Site(unsigned int xp, unsigned int yp, unsigned int zp, unsigned int xs,
+           unsigned int ys, unsigned int zs, unsigned int neigh,
+           ChemicalSystem *csys, const bool verbose) {
+  x_ = y_ = z_ = 0;
+  id_ = 0;
+  nb_.clear();
+  stressFreeVolume_ = trueVolume_ = 1.0;
+  damage_ = false;
+  x_ = xp;
+  y_ = yp;
+  z_ = zp;
 
-    #ifdef DEBUG
-        verbose_ = true;
-    #else
-        verbose_ = verbose;
-    #endif
+#ifdef DEBUG
+  verbose_ = true;
+#else
+  verbose_ = verbose;
+#endif
 
-    dissolution_.clear();
-    growth_.clear();
+  dissolution_.clear();
+  growth_.clear();
 
-    id_ = (unsigned long int)(x_ + (xs * y_) + ((xs * ys) * z_));
+  id_ = (unsigned int)(x_ + (xs * y_) + ((xs * ys) * z_));
 
-    if (neigh > 0) nb_.resize(neigh,0);
+  if (neigh > 0)
+    nb_.resize(neigh, 0);
 
-    chemSys_ = csys;
+  chemSys_ = csys;
 
-    expstrain_ = 0.0;
+  expstrain_ = 0.0;
 }
 
-void Site::calcWmc(void)
-{
-    wmc_ = chemSys_->getMicroPhasePorosity(getMicroPhaseId());
-    for (unsigned int i = 0; i < nb_.size(); i++) {
-        wmc_ += chemSys_->getMicroPhasePorosity(nb_[i]->getMicroPhaseId());
-    }
+void Site::calcWmc(void) {
+  wmc_ = chemSys_->getMicroPhasePorosity(getMicroPhaseId());
+  for (unsigned int i = 0; i < nb_.size(); i++) {
+    wmc_ += chemSys_->getMicroPhasePorosity(nb_[i]->getMicroPhaseId());
+  }
 }
